Added bonus command to add (possibly negative) bonus for a user

+++ src/state/status.rs
@@ -53,7 +53,7 @@ impl From<Option<usize>> for Score {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Ranking {
     pub max_score: usize,
-    pub scores: Vec<(String,usize)>
+    pub scores: Vec<(String,i32)>
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
